refactor UI project out of Foundation and add enhanced ReadModel support (#35)
Factor out UI project
Cleanup some Eventstore naming and classes in foundation
Add checkpoint and multi-stream support to readmodel base
increase overall code coverage

+++ src/ReactiveDomain.Foundation/StreamStore/SynchronizableStreamListener.cs
@@ -41,7 +41,7 @@ namespace ReactiveDomain.Foundation
             base.GotEvent(@event);
         }
 
-        public override void Start(string streamName, int? checkpoint = null, bool waitUntilLive = false, int millisecondsTimeout = 1000)
+        public override void Start(string streamName, long? checkpoint = null, bool waitUntilLive = false, int millisecondsTimeout = 1000)
         {
             if (Sync) {
                 SyncQueue?.Start();
